# backend/api/gen/go/admin/service/v1/i_page_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_page.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-cms/api/gen/go/content/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPageServiceCreate = "/admin.service.v1.PageService/Create"
const OperationPageServiceDelete = "/admin.service.v1.PageService/Delete"
const OperationPageServiceGet = "/admin.service.v1.PageService/Get"
const OperationPageServiceList = "/admin.service.v1.PageService/List"
const OperationPageServiceUpdate = "/admin.service.v1.PageService/Update"

type PageServiceHTTPServer interface {
	// Create 创建页面
	Create(context.Context, *v11.CreatePageRequest) (*v11.Page, error)
	// Delete 删除页面
	Delete(context.Context, *v11.DeletePageRequest) (*emptypb.Empty, error)
	// Get 获取页面数据
	Get(context.Context, *v11.GetPageRequest) (*v11.Page, error)
	// List 获取页面列表
	List(context.Context, *v1.PagingRequest) (*v11.ListPageResponse, error)
	// Update 更新页面
	Update(context.Context, *v11.UpdatePageRequest) (*v11.Page, error)
}

func RegisterPageServiceHTTPServer(s *http.Server, srv PageServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/pages", _PageService_List18_HTTP_Handler(srv))
	r.GET("/admin/v1/pages/{id}", _PageService_Get18_HTTP_Handler(srv))
	r.POST("/admin/v1/pages", _PageService_Create14_HTTP_Handler(srv))
	r.PUT("/admin/v1/pages/{id}", _PageService_Update14_HTTP_Handler(srv))
	r.DELETE("/admin/v1/pages/{id}", _PageService_Delete14_HTTP_Handler(srv))
}

func _PageService_List18_HTTP_Handler(srv PageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPageServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPageResponse)
		return ctx.Result(200, reply)
	}
}

func _PageService_Get18_HTTP_Handler(srv PageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPageServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Page)
		return ctx.Result(200, reply)
	}
}

func _PageService_Create14_HTTP_Handler(srv PageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreatePageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPageServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreatePageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Page)
		return ctx.Result(200, reply)
	}
}

func _PageService_Update14_HTTP_Handler(srv PageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdatePageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPageServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdatePageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Page)
		return ctx.Result(200, reply)
	}
}

func _PageService_Delete14_HTTP_Handler(srv PageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeletePageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPageServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeletePageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type PageServiceHTTPClient interface {
	// Create 创建页面
	Create(ctx context.Context, req *v11.CreatePageRequest, opts ...http.CallOption) (rsp *v11.Page, err error)
	// Delete 删除页面
	Delete(ctx context.Context, req *v11.DeletePageRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 获取页面数据
	Get(ctx context.Context, req *v11.GetPageRequest, opts ...http.CallOption) (rsp *v11.Page, err error)
	// List 获取页面列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPageResponse, err error)
	// Update 更新页面
	Update(ctx context.Context, req *v11.UpdatePageRequest, opts ...http.CallOption) (rsp *v11.Page, err error)
}

type PageServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPageServiceHTTPClient(client *http.Client) PageServiceHTTPClient {
	return &PageServiceHTTPClientImpl{client}
}

// Create 创建页面
func (c *PageServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreatePageRequest, opts ...http.CallOption) (*v11.Page, error) {
	var out v11.Page
	pattern := "/admin/v1/pages"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPageServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除页面
func (c *PageServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeletePageRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/pages/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPageServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 获取页面数据
func (c *PageServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPageRequest, opts ...http.CallOption) (*v11.Page, error) {
	var out v11.Page
	pattern := "/admin/v1/pages/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPageServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 获取页面列表
func (c *PageServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPageResponse, error) {
	var out v11.ListPageResponse
	pattern := "/admin/v1/pages"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPageServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新页面
func (c *PageServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdatePageRequest, opts ...http.CallOption) (*v11.Page, error) {
	var out v11.Page
	pattern := "/admin/v1/pages/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPageServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}