# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-cms
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-cms/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/pages:
        get:
            tags:
                - PageService
            description: 获取页面列表
            operationId: PageService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPageResponse'
        post:
            tags:
                - PageService
            description: 创建页面
            operationId: PageService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Page'
    /app/v1/pages/{id}:
        get:
            tags:
                - PageService
            description: 获取页面数据
            operationId: PageService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: slug
                  in: query
                  schema:
                    type: string
                - name: locale
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Page'
        put:
            tags:
                - PageService
            description: 更新页面
            operationId: PageService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Page'
        delete:
            tags:
                - PageService
            description: 删除页面
            operationId: PageService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreatePageRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Page'
            description: 请求 - 创建页面
...
        ListPageResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Page'
                total:
                    type: string
            description: 回应 - 页面列表
...
        Page:
            type: object
            properties:
                id:
                    type: integer
                    description: 页面ID
                    format: uint32
                status:
                    enum:
                        - PAGE_STATUS_UNSPECIFIED
                        - PAGE_STATUS_DRAFT
                        - PAGE_STATUS_PUBLISHED
                        - PAGE_STATUS_ARCHIVED
                    type: string
                    description: 页面状态
                    format: enum
                type:
                    enum:
                        - PAGE_TYPE_UNSPECIFIED
                        - PAGE_TYPE_DEFAULT
                        - PAGE_TYPE_HOME
                        - PAGE_TYPE_ERROR_404
                        - PAGE_TYPE_ERROR_500
                        - PAGE_TYPE_CUSTOM
                    type: string
                    description: 页面类型（首页/404/普通页面等，影响路由和渲染逻辑）
                    format: enum
                editorType:
                    enum:
                        - EDITOR_TYPE_UNSPECIFIED
                        - EDITOR_TYPE_MARKDOWN
                        - EDITOR_TYPE_RICH_TEXT
                        - EDITOR_TYPE_PLAIN_TEXT
                        - EDITOR_TYPE_CODE
                        - EDITOR_TYPE_JSON_BLOCK
                        - EDITOR_TYPE_VISUAL_BUILDER
                    type: string
                    description: 编辑器类型
                    format: enum
                slug:
                    type: string
                    description: 页面唯一标识（默认语言，URL 友好，决定主路由）
                authorId:
                    type: integer
                    description: 页面作者ID
                    format: uint32
                authorName:
                    type: string
                    description: 页面作者名称
                disallowComment:
                    type: boolean
                    description: 是否禁止评论（默认 true，页面通常关闭评论）
                redirectUrl:
                    type: string
                    description: 重定向 URL
                showInNavigation:
                    type: boolean
                    description: 是否在主导航中显示（用于自动生成菜单）
                sortOrder:
                    type: integer
                    description: 导航排序（数值越小越靠前）
                    format: uint32
                template:
                    type: string
                    description: 页面模板名称（如 'about-us', 'contact'，覆盖全局默认模板）
                isCustomTemplate:
                    type: boolean
                    description: 是否使用自定义模板代码（而非预设模板）
                visits:
                    type: integer
                    description: 页面访问次数
                    format: uint32
                customFields:
                    type: object
                    additionalProperties:
                        type: string
                    description: 自定义字段，键值对形式，便于扩展
                customHead:
                    type: string
                    description: 自定义 head 代码（如额外的 meta 标签或样式，直接插入到 <head> 中）
                customFoot:
                    type: string
                    description: 自定义 foot 代码（如额外的脚本，直接插入到 </body> 前）
                translations:
                    type: array
                    items:
                        $ref: '#/components/schemas/PageTranslation'
                    description: 多语言翻译列表
                availableLanguages:
                    example: ["zh-CN", "en-US", "ja-JP"]
                    type: array
                    items:
                        type: string
                    description: 可用的语言代码列表（快速查询，避免遍历 translations）
                parentId:
                    type: integer
                    description: 父页面ID（0 表示顶级页面，用于构建站点树形结构）
                    format: uint32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Page'
                    description: 子节点树
                depth:
                    type: integer
                    description: 页面层级深度（0=顶级，1=二级，用于面包屑导航）
                    format: int32
                path:
                    type: string
                    description: 物化路径（如 '1/5/23'，便于层级查询）
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 页面
        PageTranslation:
            type: object
            properties:
                id:
                    type: integer
                    description: 翻译记录ID
                    format: uint32
                pageId:
                    type: integer
                    description: 关联的页面ID
                    format: uint32
                languageCode:
                    type: string
                    description: 语言代码（ISO 639-1 标准）
                title:
                    type: string
                    description: 页面标题
                slug:
                    type: string
                    description: 语言特定的 slug（覆盖主表 slug，用于多语言路由，如中文用拼音）
                summary:
                    type: string
                    description: 页面摘要（用于 meta description）
                content:
                    type: string
                    description: 页面内容
                originalContent:
                    type: string
                    description: 原始内容（未渲染前）
                thumbnail:
                    type: string
                    description: 缩略图
                coverImage:
                    type: string
                    description: 封面图
                wordCount:
                    type: integer
                    description: 当前语言版本的字数（中文按字符数，英文按单词数）
                    format: uint32
                fullPath:
                    type: string
                    description: 完整路径（如 /zh-CN/about-us），首页应为 '/' 或 '/zh-CN/'
                metaKeywords:
                    type: string
                    description: SEO 关键词
                metaDescription:
                    type: string
                    description: SEO 描述
                seoTitle:
                    type: string
                    description: SEO 标题（覆盖默认 title）
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 页面翻译
...
        UpdatePageRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Page'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 请求 - 更新页面
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...